# perf: Remove allocations from image-renderer