--- a/src/engine/Graphics/Context/state-stack.ts
+++ b/src/engine/Graphics/Context/state-stack.ts
@@ -3,8 +3,8 @@ import { ExcaliburGraphicsContextState } from './ExcaliburGraphicsContext';
 import { Material } from './material';
 
 export class StateStack {
+  public current: ExcaliburGraphicsContextState = this._getDefaultState();
   private _states: ExcaliburGraphicsContextState[] = [];
-  private _currentState: ExcaliburGraphicsContextState = this._getDefaultState();
 
   private _getDefaultState() {
     return {
@@ -17,27 +17,19 @@ export class StateStack {
 
   private _cloneState() {
     return {
-      opacity: this._currentState.opacity,
-      z: this._currentState.z,
-      tint: this._currentState.tint.clone(),
-      material: this._currentState.material // TODO is this going to cause problems when cloning
+      opacity: this.current.opacity,
+      z: this.current.z,
+      tint: this.current.tint.clone(),
+      material: this.current.material // TODO is this going to cause problems when cloning
     };
   }
 
   public save(): void {
-    this._states.push(this._currentState);
-    this._currentState = this._cloneState();
+    this._states.push(this.current);
+    this.current = this._cloneState();
   }
 
   public restore(): void {
-    this._currentState = this._states.pop();
-  }
-
-  public get current(): ExcaliburGraphicsContextState {
-    return this._currentState;
-  }
-
-  public set current(val: ExcaliburGraphicsContextState) {
-    this._currentState = val;
+    this.current = this._states.pop();
   }
 }