… CHANGELOG.md
… src/spec/VertexLayoutSpec.ts