<!--
Hi, and thanks for contributing to Excalibur!
Before you go any further, please read our contributing guide: https://github.com/excaliburjs/Excalibur/blob/main/.github/CONTRIBUTING.md
especially the "Submitting Changes" section:
https://github.com/excaliburjs/Excalibur/blob/main/.github/CONTRIBUTING.md#submitting-changes
---
A quick summary checklist is included below for convenience:
-->

===:clipboard: PR Checklist :clipboard:===

- [ ] :pushpin: issue exists in github for these changes
- [ ] :microscope: existing tests still pass
- [x] :see_no_evil: code conforms to the [style guide](https://github.com/excaliburjs/Excalibur/blob/main/STYLEGUIDE.md)
- [x] :triangular_ruler: new tests written and passing / old tests updated with new scenario(s)
- [x] :page_facing_up: changelog entry added (or not needed)

==================

See excalibur bunnymark https://github.com/excaliburjs/excalibur-bunnymark

This PR removes allocations and makes several perf improvements that DOUBLES our performance on the bunnymark

* Removes all allocations from the draw image hot path
* Removes allocations from object pool
* Removes unecessary getters
* Speed up AffineMatrix clone
* Switches webgl to VAOs
* Speed up texture id lookup
* Caches native thunks to width/height

## New Perf Results

 ~50fps @ 20,000 sprites.

![image](https://github.com/excaliburjs/Excalibur/assets/612071/292a5ff5-4e03-4d19-9118-a2f3455960c1)

~30fps @ 35000 sprites

![image](https://github.com/excaliburjs/Excalibur/assets/612071/8abbb13b-cfcf-4b38-9d89-dc5def385d11)


## Previous Perf Results on Excalibur main 


~25fps @ 20000 sprites

![image](https://github.com/excaliburjs/Excalibur/assets/612071/afe0bab7-b2ee-4edf-a3c0-d9ede0325cb5)


~15fps @ 35000 sprites
![image](https://github.com/excaliburjs/Excalibur/assets/612071/391bb767-e712-494a-b442-fd7425af8637)